+++ src/utils/index.js
@@ -1,2 +1,3 @@
 export * from "./event.js";
 export * from "./price.js";
+export * from "./validate.js";
